1st Day
This time around (lol.) I'm just using objects instead of turning the Monsters into Nodes to then do things with. Didn't know that methods on objs can just be called. makes it a lot easier to just store an array of the party/storage and use autoload singletons to send signals from various Effects to the battle UI

Decent start for today. I hope i never remake this stupid thing ever again. I really want to build the game. I hope this structure will be easier for me to work with and not buckle under the weight of my inexperience and naivete