Basically done with Effect/Status refactor
Very happy with the new system. Was able to use deepseek to convert all of my old effects/statuses into a new component based system that can build seemingly any effect or status, with adding more very very easy.

New scripts will only have to be made when I come up with a radically different effect that doesn't fit into the current framework.

An example would be a stacking component, or maybe a weather/terrain setter. One that I envision is a "Flag-Setter" component that has two string fields and just executes Battlemanager.flagX = flagY

This could be used to set terrain/weather and that type of thing. Maybe it also has a third flag of number of turns to refresh for and fourth flag for whether it turns off at the end